ml-analysis/assets/project-template/requirements.txt
# Core ML
scikit-learn>=1.4
numpy>=1.26
pandas>=2.1

# Gradient boosting (preferred)
lightgbm>=4.3
xgboost>=2.0

# Hyperparameter tuning
optuna>=3.5

# Explainability
shap>=0.44

# Data quality
great-expectations>=0.18

# Serving
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5

# Experiment tracking
mlflow>=2.11

# Drift detection
evidently>=0.4

# Utilities
joblib>=1.3
pyyaml>=6.0
scipy>=1.12

# Notebooks
jupyter>=1.0
matplotlib>=3.8
seaborn>=0.13

# Optional: large-scale
# dask[dataframe]>=2024.1
# pyspark>=3.5
# polars>=0.20
